Refactor content types into dedicated modules
