+++ packages/spectrum-ts/src/content/text.ts
@@ -1,0 +1,13 @@
+import z from "zod";
+import type { ContentBuilder } from "./types";
+
+export const textSchema = z.object({
+  type: z.literal("text"),
+  text: z.string().nonempty(),
+});
+
+export function text(text: string): ContentBuilder {
+  return {
+    build: () => Promise.resolve({ type: "text", text }),
+  };
+}
