+++ packages/spectrum-ts/src/platform/types.ts
@@ -1,6 +1,6 @@
 import type { Fn, Pipe, Tuples } from "hotscript";
 import type z from "zod";
-import type { Content } from "../types/content";
+import type { Content } from "../content/types";
 import type { Message } from "../types/message";
 import type { Space } from "../types/space";
 import type { User } from "../types/user";
